Restructure consensus data structure folders
