+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["node", "braidpool-primitives"]
+members = ["node"]
 resolver = "3"
