--- braidpool-primitives/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "braidpool-primitives"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-hex = "0.4.3"
-num-bigint = "0.4.6"
-ordered-float = "5.0.0"
-transitive = "1.1.0"
-bitcoin = { git = "https://github.com/braidpool/rust-bitcoin.git" ,features=["serde"]}
-serde = {version = "1.0.219", features=["derive"]}
-serde_json = "1.0.140"
-secp256k1 = "0.30.0"
-
-[lib]
-name = "braidpool_primitives"
-path = "src/lib.rs"
