--- braidpool-primitives/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod bead;
-pub mod braid;
-pub mod utils;
