--- braidpool-primitives/tests/bead_tests.rs
@@ -1,1 +1,0 @@
-// Integration tests for public functions!
