--- braidpool-primitives/tests/braid_tests.rs
@@ -1,1 +1,0 @@
-// Integration tests for the braid
