+++ node/Cargo.toml
@@ -3,10 +3,6 @@ name = "node"
 version = "0.1.0"
 edition = "2021"
 
-[[test]]
-name = "braid_tests"
-path = "../tests/braid_tests.rs"
-
 [dependencies]
 tokio = { version = "1", features = ["signal","rt", "net", "macros", "rt-multi-thread", "tracing", "io-util"] }
 serde = { version = "^1.0", features = ["derive"] }
@@ -18,7 +14,7 @@ tracing-subscriber = { version = "0.3.1", default-features = false, features = [
 futures = { version = "0.3.28", features = ["thread-pool"]}
 bytes = "1.4.0"
 flexbuffers = { version = "25.2", features = ["deserialize_human_readable", "serialize_human_readable"] }
-serde_json = "1.0.10"
+serde_json = "1.0.140"
 sqlite = { version = "0.36" }
 clap = { version = "4.4.2", features = ["derive", "string"] }
 bitcoin = { git = "https://github.com/braidpool/rust-bitcoin.git", branch = "cpunet" }
@@ -38,4 +34,8 @@ bimap = "0.6.3"
 ordered-float = "3.7.0"
 lazy_static = "1.4.0"
 regex = "1.11.1"
+hex = "0.4.3"
+num-bigint = "0.4.6"
+transitive = "1.1.0"
+secp256k1 = "0.30.0"
 libp2p = { version = "0.55.0", features = ["identify", "quic", "tokio","ping","macros"] }
