+++ node/src/lib.rs
@@ -1,1 +1,6 @@
+pub mod bead;
 pub mod braid;
+pub mod braid_functions;
+pub mod committed_metadata;
+pub mod uncommitted_metadata;
+pub mod utils;
