+++ node/src/utils/mod.rs
@@ -1,0 +1,27 @@
+// Bitcoin Imports
+use crate::bead::Bead;
+use ::bitcoin::BlockHash;
+// Standard Imports
+
+pub mod test_utils;
+
+// External Type Aliases
+pub type BeadHash = BlockHash;
+pub type Byte = u8;
+pub type Bytes = Vec<Byte>;
+
+// Internal Type Aliases
+pub(crate) type Relatives = HashSet<BeadHash>;
+
+// Error Definitions
+use std::collections::HashSet;
+
+pub(crate) fn hashset_to_vec_deterministic(hashset: &HashSet<BeadHash>) -> Vec<BeadHash> {
+    let mut vec: Vec<BeadHash> = hashset.iter().cloned().collect();
+    vec.sort();
+    vec
+}
+
+pub(crate) fn vec_to_hashset(vec: Vec<BeadHash>) -> HashSet<BeadHash> {
+    vec.iter().cloned().collect()
+}
